+++ src/base/widget.h
@@ -1901,6 +1901,7 @@ bool_t widget_is_instance_of(widget_t* widget, const widget_vtable_t* vt);
 TK_EXTERN_VTABLE(widget);
 
 /*public for test*/
+ret_t widget_move_focus(widget_t* widget, bool_t next);
 locale_info_t* widget_get_locale_info(widget_t* widget);
 image_manager_t* widget_get_image_manager(widget_t* widget);
 assets_manager_t* widget_get_assets_manager(widget_t* widget);