+++ docs/changes.md
@@ -3,6 +3,7 @@
   * 修改edit在scrollview中的问题。
   * resize后设置need_relayout_children标志(感谢朝泽提供补丁)。
   * 修改slide menu中find target的bug(感谢朝泽提供补丁)。
+  * move focus时跳过不可见和不可用的widget。
    
 * 2019/05/22
   * 完善事件处理流程，修改slide menu放在slide view中的问题。